Nintendo logo!!
Connected PPU to ebiten renderer, and it basically works! FPS is around
20 with the renderer, but that's surprisingly good for 0 optimizations!
There are some black bars on the screen, probably a bug in the
background tile writing code.

Not sure what the first optimization step is, it's possible the videoout
channel is blocking. I'll need to profile to see what's really going on.
But 20fps first try is something I'm VERY happy with!!